+++ otail-web/src/App.tsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'
 import Layout from './layout'
-import Sampling from './smapling/page'
-import OtelConfig from './config/page'
-import Agents from './agents/page'
+import Sampling from './pages/sampling'
+import OtelConfig from './pages/config'
+import Agents from './pages/agents'
 import { ThemeProvider } from "@/hooks/use-theme"
 
 function App() {
